undid introduction of tests/__init__.py, instead made a testing.py module for handy functions used by testing and perhaps average user
