--- hera_pspec/tests/__init__.py
@@ -1,2 +1,0 @@
-""" __init__.py for hera_pspec/tests """
-from noise_sim import noise_sim
